update default assistance request service

+++ engelsizgonuller/engelsizgonuller/src/main/java/dev/melis/engelsizgonuller/services/assistancerequest/DefaultAssistanceRequestService.java
@@ -1,9 +1,11 @@
 package dev.melis.engelsizgonuller.services.assistancerequest;
 
-import dev.melis.engelsizgonuller.business.result.CreationResult;
+import dev.melis.engelsizgonuller.support.result.CreationResult;
 import dev.melis.engelsizgonuller.repository.AssistanceRequestRepository;
 import dev.melis.engelsizgonuller.services.model.helpassistance.AssistanceRequests;
 import dev.melis.engelsizgonuller.services.model.helpassistance.RequestType;
+import dev.melis.engelsizgonuller.support.result.OperationFailureReason;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -20,40 +22,49 @@ public class DefaultAssistanceRequestService implements AssistanceRequestService
         this.assistanceRequestRepository = assistanceRequestRepository;
     }
     @Override
-    public CreationResult createAssistanceRequest(AssistanceRequestServiceRequest assistanceRequests) {
+    public CreationResult createAssistanceRequest(AssistanceRequestServiceRequest serviceRequest) {
+        Optional<AssistanceRequests> requestOptional=assistanceRequestRepository.findByRequestHeaderAndUser(serviceRequest.getRequestHeader(),serviceRequest.getUser());
+        if(requestOptional.isPresent()){
+            return CreationResult.failure(OperationFailureReason.CONFLICT,"Request already exist");
+        }
         var assistanceRequest= new AssistanceRequests();
-        assistanceRequest.setUser(assistanceRequests.getUser());
-        assistanceRequest.setRequestContent(assistanceRequests.getRequestContent());
+        assistanceRequest.setUser(serviceRequest.getUser());
+        assistanceRequest.setRequestHeader(serviceRequest.getRequestHeader());
+        assistanceRequest.setRequestContent(serviceRequest.getRequestContent());
         assistanceRequest.setRequestDate(LocalDate.now());
-        assistanceRequest.setCategory(assistanceRequest.getCategory());
-        assistanceRequest.setRequestDeadline(assistanceRequests.getRequestDeadline());
-        assistanceRequest.setFulFilled(assistanceRequests.isFulFilled());
-        if(assistanceRequests.getRequestType()==RequestType.VOLUNTEER_HELP){
+        assistanceRequest.setCategory(serviceRequest.getCategory());
+        assistanceRequest.setRequestDeadline(serviceRequest.getRequestDeadline());
+        assistanceRequest.setFulFilled(serviceRequest.isFulFilled());
+        if(serviceRequest.getRequestType()==RequestType.VOLUNTEER_HELP){
             assistanceRequest.setRequestType(RequestType.VOLUNTEER_HELP);
         }
-        assistanceRequest.setRequestType(RequestType.DISABLED_REQUEST);
+        else{
+            assistanceRequest.setRequestType(RequestType.DISABLED_REQUEST);
+        }
         assistanceRequestRepository.save(assistanceRequest);
         return CreationResult.success();
     }
 
     @Override
-    public boolean delete(Long id) {
+    public boolean delete(Long id,Long userId) {
         Optional<AssistanceRequests> existsRequest= assistanceRequestRepository.findById(id);
         if(existsRequest.isPresent()){
+            if(existsRequest.get().getUser().getUserId()!=userId){
+                return false;
+            }
             assistanceRequestRepository.deleteById(id);
             return true;
         }
         return false;
     }
-
-    @Override
-    public List<AssistanceRequests> listRequestWithCategory(Long id) {
-        List<AssistanceRequests> assistanceRequestsList= (List<AssistanceRequests>) assistanceRequestRepository.findByCategory(id);
-        if(assistanceRequestsList!=null && !assistanceRequestsList.isEmpty()){
-            return assistanceRequestsList;
+        @Override
+        public List<AssistanceRequests> listRequestWithCategory(Long id) {
+            List<AssistanceRequests> assistanceRequestsList= (List<AssistanceRequests>) assistanceRequestRepository.findAllByCategoryCategoryId(id);
+            if(assistanceRequestsList!=null && !assistanceRequestsList.isEmpty()){
+                return assistanceRequestsList;
+            }
+            return new ArrayList<>();
         }
-        return new ArrayList<>();
-    }
 
     @Override
     public List<AssistanceRequests> listRequestWithRequestType(RequestType requestType) {
